Relocate relapath and reduce theoretical complexity of index filters
