+++ crates/but-workspace/src/discard/mod.rs
@@ -34,6 +34,21 @@ impl From<DiscardSpec> for DiffSpec {
     }
 }
 
+pub(crate) trait RelaPath {
+    fn rela_path(&self) -> &bstr::BStr;
+}
+
+impl RelaPath for gix::diff::index::ChangeRef<'_, '_> {
+    fn rela_path(&self) -> &bstr::BStr {
+        match self {
+            gix::diff::index::ChangeRef::Addition { location, .. }
+            | gix::diff::index::ChangeRef::Modification { location, .. }
+            | gix::diff::index::ChangeRef::Rewrite { location, .. }
+            | gix::diff::index::ChangeRef::Deletion { location, .. } => location,
+        }
+    }
+}
+
 pub(super) mod function;
 #[allow(missing_docs)]
 pub mod ui {
